+++ lib/src/converters/converters.dart
@@ -5,7 +5,7 @@ import 'dart:typed_data';
 import 'package:flutter/material.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:layrz_icons/layrz_icons.dart';
-import 'package:layrz_models/utils/utils.dart';
+import 'package:layrz_models/src/utils/utils.dart';
 
 part 'src/color.dart';
 part 'src/duration.dart';