+++ lib/src/tenvio/tenvio.dart
@@ -4,7 +4,6 @@ import 'package:flutter/material.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:layrz_icons/layrz_icons.dart';
 import 'package:layrz_models/layrz_models.dart';
-import 'package:layrz_models/utils/utils.dart';
 
 part 'tenvio.freezed.dart';
 part 'tenvio.g.dart';